refactor and move to helper
